[Y26W2-386] fix: 비교표 조회 시 서비스 레이어에서 DTO반환 - 컨트롤러 레이어에서 트랜잭션 분리
