+++ src/main/resources/config/postgresql/db-dev.yml
@@ -6,6 +6,7 @@ spring:
     driver-class-name: org.postgresql.Driver
     url: jdbc:postgresql://${DB_PRIVATE_IP}:5432/${TEST_DB_NAME}
   jpa:
+    open-in-view: false
     database-platform: org.hibernate.dialect.PostgreSQLDialect
     hibernate:
       ddl-auto: validate
